HP 자료형 수정 및 기타 코드 수정
- 몬스터/ 플레이어 hp 저장 변수 float 자료형으로 변경

- 서버 lua_functions.cpp  코드 복구

- ui  SetEnemy후에 Attach

+++ Client/Revive/src/object/actor/character/skeleton_soldier.cpp
@@ -70,7 +70,7 @@ namespace revive
 				const auto& player = std::dynamic_pointer_cast<DefaultPlayer>(other_actor);
 				if (player != nullptr)
 				{
-					int player_hp = player->GetHP();
+					float player_hp = player->GetHP();
 					if (player_hp > 0)
 					{
 							//플레이어 피격했다고 서버에 알리기
@@ -82,7 +82,7 @@ namespace revive
 					const auto& base = std::dynamic_pointer_cast<Base>(other_actor);
 					if (base != nullptr)
 					{
-						int base_hp = base->GetHP();
+						float base_hp = base->GetHP();
 						if (base_hp > 0)
 						{
 							//기지 피격했다고 서버에 알리기
@@ -151,7 +151,7 @@ namespace revive
 					{
 						m_target_position = other_actor->GetPosition() + Vec3{ 0.f,200.f,0.f };
 
-						int base_hp = base->GetHP();
+						float base_hp = base->GetHP();
 						if (base_hp > 0)
 							Attack();
 					}
